+++ trilby-cli/CHANGELOG.md
@@ -5,6 +5,8 @@
 * Introduce [CalVer](https://calver.org/) versioning (YY.MM.MICRO)
 * Add `--version` flag
 * Replace `--verbosity` with `--quiet`, `--verbose`, and `--debug`
+* Drop direct dependency on `terminal`
+
 ## 0.1.0.0 -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.